test: keep root package file count within baseline
